working servos and motor

+++ TimerA.h
@@ -5,5 +5,5 @@
 // Functions
 int TIMER_A0_PWM_Init(uint16_t period, double percentDutyCycle, uint16_t pin);
 void TIMER_A0_PWM_DutyCycle(double percentDutyCycle, uint16_t pin);
-int TIMER_A2_PWM_Init(uint16_t period, double percentDutyCycle, uint16_t pin);
+int TIMER_A2_PWM_Init(uint32_t period, double percentDutyCycle, uint16_t pin);
 void TIMER_A2_PWM_DutyCycle(double percentDutyCycle, uint16_t pin);
